requirements.txt
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.12.0
Flask==2.2.2
fonttools==4.39.4
gunicorn==20.1.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywatchman==1.4.1
pyzmq==25.0.2
redis==4.5.5
requests==2.30.0
six==1.16.0
spotipy==2.23.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.6
Werkzeug==2.3.4
zope.interface==6.0
